~ share*->get_wires() vector reference instead of copy
The intuitively expected behavior of get_wires() is to return a reference to the vector of wires for modifying the inside of the share. However, the current implementation returns a copy of the vector and all the changes to this vector can be propagated to the circuit only via creating a new share.